odb: absorb threeDBloxValidator in Checker
