+++ src/odb/src/3dblox/checker.h
@@ -4,11 +4,14 @@
 #pragma once
 
 #include "odb/db.h"
+#include "odb/geom.h"
+#include "unfoldedModel.h"
+#include "utl/Logger.h"
 
 namespace odb {
 class dbChip;
+class dbMarkerCategory;
 
-// Facade for backward compatibility and ease of use
 class Checker
 {
  public:
@@ -17,6 +20,23 @@ class Checker
   void check(odb::dbChip* chip);
 
  private:
+  void checkFloatingChips(const UnfoldedModel& model,
+                          dbMarkerCategory* category);
+  void checkOverlappingChips(const UnfoldedModel& model,
+                             dbMarkerCategory* category);
+  void checkConnectionRegions(const UnfoldedModel& model,
+                              dbChip* chip,
+                              dbMarkerCategory* category);
+  void checkBumpPhysicalAlignment(const UnfoldedModel& model,
+                                  dbMarkerCategory* category);
+  void checkNetConnectivity(const UnfoldedModel& model,
+                            dbChip* chip,
+                            dbMarkerCategory* category);
+
+  bool isOverlapFullyInConnections(const UnfoldedChip* chip1,
+                                   const UnfoldedChip* chip2,
+                                   const Cuboid& overlap) const;
+
   utl::Logger* logger_;
 };
